+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## 5.1.0 (unreleased)
 
-* **dashboards**: bump revision of all dashboards due to the internal plugin IDs change (see below)
+* **redis**: validate base URL
 * **vector,bpftrace**: increase data source settings form column to prevent line wrap
+* **dashboards**: bump revision of all dashboards due to the internal plugin IDs change (see below)
 * **build**: update to Grafana v9.0.9 and sync minimum Grafana version requirement
 * **docs**: update version compatibility table