+++ CHANGELOG.md
@@ -5,8 +5,9 @@
 * **redis**: validate base URL
 * **vector,bpftrace**: increase data source settings form column to prevent line wrap
 * **dashboards**: bump revision of all dashboards due to the internal plugin IDs change (see below)
-* **build**: update to Grafana v9.0.9 and sync minimum Grafana version requirement
+* **build**: update dependencies, update to Grafana v9.0.9 and sync minimum Grafana version requirement
 * **docs**: update version compatibility table
+* **ci**: upgrade cypress
 
 
 ## 5.0.0 (2022-06-30)